requirements.txt
# AI Video Watcher - Requirements
# Core Streamlit and video processing
streamlit>=1.28.0
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# OpenAI integration
openai>=1.0.0

# Audio processing and transcription
openai-whisper>=20231117
pyannote.audio>=3.1.1
torch>=2.0.0
torchaudio>=2.0.0
resampy
soundfile

# Audio analysis
tensorflow
tensorflow-hub

# Report generation
reportlab

# Note: FFmpeg system binary must be installed separately
# Ubuntu/Debian: sudo apt install ffmpeg
# MacOS: brew install ffmpeg
# Windows: Download from https://ffmpeg.org/
ffmpeg
reportlab
openai
psycopg2-binary
pyannote.audio